+++ go/normalize.go
@@ -42,5 +42,6 @@ func isTrimWhitespace(r rune) bool {
 	return r == 0x9 || (r >= 0xa && r <= 0xd) || r == 0x20 || r == 0xa0 || r == 0x1680 || (r >= 0x2000 && r <= 0x200a) || r == 0x2028 || r == 0x2029 || r == 0x202f || r == 0x205f || r == 0x3000 || r == 0xfeff
 }
 
-func trimWhitespace(s string) string     { return strings.TrimFunc(s, isTrimWhitespace) }
-func trimLeftWhitespace(s string) string { return strings.TrimLeftFunc(s, isTrimWhitespace) }
+func trimWhitespace(s string) string      { return strings.TrimFunc(s, isTrimWhitespace) }
+func trimLeftWhitespace(s string) string  { return strings.TrimLeftFunc(s, isTrimWhitespace) }
+func trimRightWhitespace(s string) string { return strings.TrimRightFunc(s, isTrimWhitespace) }